+++ src/main/java/nestorcicardini/D1/entities/Product.java
@@ -12,9 +12,9 @@ import lombok.ToString;
 public abstract class Product {
 	private String name;
 	private double price;
-	private int calories;
+	private double calories;
 
-	public Product(String name, double price, int calories) {
+	public Product(String name, double price, double calories) {
 		super();
 		this.name = name;
 		this.price = price;
